scope contacts to a department
This is the first step to making this more generic, we scope all the
contacts to hmrc.

Also removed the hmrc colors and some references.

Still need to rename the app, modules, remove hmrc hard coding in the
designs, update partial names, make the admin interface have a
department filter.

Ran into issues with ActiveYaml and Decent Exposure when using scopes,
so made Department an ActiveRecord object