Use form tag rather than form helper
The form helper adds a UTF8 tick input which is shown in the URL bar.
This can cause lots of problems when people try and share URLs. The
reason the tick is added to trick older versions of IE into sending UTF8
encoded post data. As this is a GET request we don't need to trick it.