Simplify contacts factory
Although it's not a validation, we would always expect a contact to have
an organisation. The :organisation trait was never used and also broken
anyway (I'm assuming because it had the same name as an association on
the model), and the little dance to set all the contact's groups to be
owned by the same org is unnecessary and likely to cause confusion if
someone explicitly sets a gorup that isn't owned by the same org.