Configure asset host in development mode.
When run under `govuk_setenv`, this configures the asset_host to point
at the assets-origin on the dev VM.  This will give us better dev-prod
parity, and enable running this behind the router in dev.

This also removes the override of the assets prefix in development.
There shouldn't be a need to override this.