Remove hokey JS on update action
The app was using remote (JS) form-submission to update the "more info"
attribute on the main Contact model from the various tabs. The only
reason I can think that it was doing this was so to keep the user on the
current tab (whereas doing a standard Rails non-AJAX update would
redirect the user back to the main contact tab). This replaces the
remote form submission, using a hidden parameter in the form to identify
where the user should be sent after a successful update.

This does mean that the user will be shown the contacts#edit page if
there are validation errors, but given that there are no validations on
any of the "more info" attributes, this should never happen.