Pin `select2-rails` to a specific version.
We shouldn't include dependencies on a third-party GitHub repo without
locking it down to a specific tag or commit. Besides, there's a released
gem available.

This is the closest version after the commit in `Gemfile.lock`. Because
the maintainer doesn't follow SemVer, we need to pin this to a single
version.