Version the `rummageable` dependency.
We shouldn't be depending on the `master` branch of a dependency,
because that can introduce backwards-incompatible changes. The commit
from `Gemfile.lock` is the release commit for version `1.0.1`, so we'll
use that as a minimum; because we're using SemVer, we only need to
restrict the version number to `1.x` to avoid backwards-incompatible
changes.