Add related contacts to the contact form.
For now, we can get away with using selecting all the contacts and
dumping them into a multi-select field, because there's only around 100
of them. If and when we get to the point of using this for more contact
details from around government, we're going to need to do something a
little smarter with this field, because:

  * the page is going to get slower and slower as we add contacts
  * it's going to get harder to find the contact you're looking for

This should be good enough for now, though.