Populate and validate content IDs on contacts.
The level of validation on UUIDs is potentially more than we need for
this app alone, but the plan is to pull that code out to use in the
content store and potentially extract as a common content ID library.

We need an `Inflector` rule set up so that Rails' auto-loader knows to
look for `UUIDValidator` rather than `UuidValidator`.

We also need to declare our own model in the migration, because the
`attr_readonly` declaration stops us populating content IDs in existing
items.