Create a Gone item when deleting contacts
Before this change, when a contact was deleted, it was left in the content
store, and so remained published on GOV.UK. This bug probably appeared when the
contacts app was split into two and moved to use the content store.

This change means that GOV.UK should display a 410 Gone page.

Ideally we would provide an additional facility to redirect deleted contacts to
another contact or another page, but this change seems far better than silently
leaving old content published.