Improve admin filtering by title and description
It was actually working, but because it did an exact match it was hard to use.

A substring match is much more helpful. It also seemed unhelpful to search
separately across two fields.

The ContactsSearch is used by the public-facing frontend too. That was working
because it was calling the "name" method which was using the substring match
code.