Revert "Fix broken frontend assets"
Contrary to my testing precompiling assets in development, Rails does try to get
the asset_host when precompiling on preview. The code depended upon the request
to do the right thing, so the best idea seems to be to revert it and ignore the
issue with respond.js and IE<9 until we split the frontend off into a finder.

This reverts commit 470bd70956e52643cc4281eea32e8c681fec1c10.