Add failing test for deleting items from search
The code for deleting items from GOV.UK site search wouldn't have worked because
it was using the slug. Rummager needs the path to the item, omitting the leading
slash.

This hasn't been an issue for contacts, because they aren't added to the site
search index.

Theoretically, it could have deleted mainstream content with slugs matching the
contact slug, but a review of the slugs used indicates this hasn't happened.