Publish a Finder for Contacts Per Org
As we want to move Contacts Index pages to a Finder, this commit
publishes a Finder every time a User updates a Contact Group.

We're doing this on the Contact Group as we only need to do it every
time a group changes as it's only regenerating the facets and publishing
the Finder again. It uses the Organisation.contact_finder_uuid that I
added in the preview commit as the UUID for the content item.

This commit includes a Publish Finder class and a Presenter for the
ContentItem. It also includes the changes to the ContactGroup to trigger
the publishing after_save.