Prevent concurrent organisation imports
We should only do one of these at a time, but the cron job runs on all (3)
servers, so can trample over itself.

I suspect this is the cause of duplicate organisations in our database - the
when one instance calls find_or_create_by(slug: slug), it finds nothing, so
creates a new organisation, but another instance has done the same thing a
fraction of a second earlier.