Add contacts to search as "Contact" documents
We are trying to replace the contact index page that is currently handled by
this application with a Finder. For finders to work, they need the things they
are listing to be of a specific document type.

Had to switch from Rummageable to GdsApi::Rummager because Rummageable doesn't
support custom document types. This also reduced the size of the dependency
list.

Removed lines that were handling Rummageable-specific exceptions. If we're going
to do that, we should do it consistently.