Register routes using publishing API
we've introduced a new `special_route` format[1] to allow us to register
unusual routes which don't have a real format associated with them.

This adds a rake task to register the HMRC contact route via that mechanism.

The intention is to run this rake task on deploy (as we do for the route
registration at present).

[1] https://github.com/alphagov/govuk-content-schemas/pull/106