Add destroy and redirect interactor for Contact
This acts as an alternative to the existing DestoryContact interactor for removing a contact form the db.  Both remove it from the db and rummager but while the existing DestroyContact one publishes a "gone" item, the new DestroyAndRedirect one publishes a redirect item pointing to a new location.

There will be no web-UI for this as we've not had lots of requests to do this, but this paves the way for it should we need to in the future.