Add contacts:replace_gone_with_redirect rake task
This is a script for the situation where the contact has already been
deleted in the app, but we'd rather there was a redirect in place than
a gone.  There's no interactor as we do not envisage having to provide
this functionality via the web-UI at any point.  We know we have to do
it at least once though and scripting it is useful due to it being a
non-trivial process to run through correctly.

We check first that:

  1. the Contact doesn't exist in the db
  2. there is an entry in content-store for it
  3. and that entry is a "gone" item

Only if all of these checks pass will we push the redirect item.