Don't check publishing_app before redirecting
We were being super cautious and making sure that the publishing_app of a
removed contact was `contacts-admin` and not publishing the redirect if it
wasn't.  However, the representation of a content-item that you get from
the content-store api doesn't include the publishing_app because
content-store provides frontend representations of items and the
publishing_app is really a backend concern.

We're fairly sure this will be ok because we don't let you jsut redirect
any path, it's a constructed path for a contact object and only
contacts-admin would publish those paths.