Create 'hmrc' organisation only if it does not exist
The README states to do the following for setting up the local database:
```
bundle exec rake db:schema:load
bundle exec rake db:seed
bundle exec rake contacts:import_hmrc DATA_FILE=db/contact-records.csv
```
This fails with a clean local install, because an organisation with the slug 'hm-revenue-customs' is being created twice (so fails with `ActiveRecord::RecordNotUnique`).
This is because we already import all organisation from Whitehall as part of the seed. To fix, we first check to see if an organisation with that slug can be found and only create it if it does not exist.