Add Publisher class to combine publish methods
To publish a content-item, we need to first send the content to draft, the links to draft and finally publish the item to production. This class handles these methods in one place. Tests have been added for this class and other classes and specs updated to use this new Publisher class.

The first class method (`self.publish`) in Publisher is for convenience with testing.

`Redirect_for_gone_item` Class is updated to make publishing calls through `Publisher` instance, therefore does not have to do any response handling