Remove ContactOrganisation and ContactGroup Presenters
As of we send organisations as part of the links hash with only a content_id. Therefore we can remove the ContactOrganisationPresenter.
We also remove the ContactGroups Presenter, as this data does not yet need to be stored in the content-store. ContactsGroups are used by the admin app only.

This should reviewed at the same time as [this](https://github.com/alphagov/govuk-content-schemas/pull/211) change to `govuk-content-schema`.