Only send the organisations link on create
Finding Things is "migrating" the tagging for all applications[1].
Migration means that publisher apps should read & write the links hash
from the publishing-api. This allows pages to be tagged by multiple
apps. For example, a Whitehall document will be taggable in Whitehall,
content-tagger and (in the future) a bulk-tagger application. Instead
of the publisher apps having knowledge of the tags in their database,
we have it centrally stored in the publishing-api.

Specifically for this application, it means that we should read & write
the `organisations` links field from the publishing-api.

Originally, we wanted to refactor this app so that the
organisation-dropdown for Contact is populated from the publishing-api
and not saved in the database (#230). This proved very difficult
because the Contact model uses the organisation to determine its base
path, among other things.

The solution we have come up with is this:

- Users can only select the organisation for the contact on creation
- After creation, we send the organisation to the publishing-api _once_

This means that if the `organisations` in the links is modified,
contacts-admin will not accidentally overwrite it - and we can start
tagging contact pages with content-tagger. We don't really change the
current behaviour, as organisations don't get edited anyway (everything
is HMRC).

Trello: https://trello.com/c/afa0Mfa4

Paired with @Davidslv

[1] https://trello.com/c/A4P3V9gK