Upgrade to rails 4.2.7.1
This addresses CVE-2016-6316 (and CVE-2016-6317 although that's not a
problem for rails 4.1.x).

As well as the rails gems we also have to upgrade friendly_id, govspeak,
json_schema, and govuk_admin_template to get new versions that explicitly
work with rails 4.2, or don't cause dependency problems.

As part of the upgrade we also update standard rails files to their 4.2
versions.  We've not replaced config/initializers/secret_token.rb with
secrets.yml though as this requires puppet changes too, and we really
want to get this security release out there quickly.