Avoid sass errors for extending with something that has no effect
Rails 4.2 uses sass-rails 5 which uses sass 3.4.  Sass 3.3 added the
following breaking change:

    Sass will now throw an error when an `@extend` that has no effect is
    used. The `!optional` flag may be used to avoid this behavior for a
    single `@extend`.

In _organisations.scss we extend a few selectors with another selector
that defines no rules (maybe it used to, or is being used for semantic
reasons).  The error message that sass generates suggests that it's
because sass can't find the selector we are trying to extend with, but
it's really because it doesn't cause any changes.  To fix this we add
a no-op vendor prefix rule `-sass-extend: keep;` that does nothing, but
means sass sees a change.

A better fix would probably be to reduce our dependency on extend and
use mixins.  Or avoid the empty extends entirely.