Catch exceptions for gone and not found items in content store
Release 32.3.0 of gds-api-adapters introduced that not found and gone items in the content store would throw exceptions which will be the default behaviour from October 2016 onwards.
We update the RedirectorForGoneContact class to handle the exceptions accordingly.

There is some uncertainty whether this class still does the right thing and should be updated to use the newer `/unpublish` endpoint