Opt in to new GDS API Adapters behaviour
[Trello card](https://trello.com/c/B5z6Dq8q/)

Some old behaviour has been deprecated in [gds-api-adapters](https://github.com/alphagov/gds-api-adapters) so we are updating this app to comply with this before it becomes the default. Firstly, a GdsApi::NotFound error is now raised when the server returns a 404 or 410.  Secondly, requests now return a hash instead of an OpenStruct.  We have now opted in to this new behaviour.

OpenStructs have been converted to hashes where necessary, and where this data is being accessed has been modified accordingly.  In particular, we have removed delegator methods from WorldLocation and replaced them with separate methods as they are no longer suitable to retrieve title and slug information.

As part of this work, we made a change to gds-api-adapters (see [PR](https://github.com/alphagov/gds-api-adapters/pull/582) to ensure a results hash is returned in ListResponse instead of an OpenStruct when apps are opted in to the new behaviour.