Migrate the HMRC contacts page to finder-frontend
This commit migrates the rendering of the HMRC contacts page at https://www.gov.uk/government/organisations/hm-revenue-customs/contact from contacts-admin to finder-frontend, since the formats are so similar.

A service and rake task are provided to publish the finder to the Publishing API. This also happens every time a contact group is added/edited/deleted in the admin interface since the contact groups are included in the JSON that is sent to the Publishing API.

Finally, contact groups are sent to rummager to be indexed so that the finder facets work.