Update airbrake to latest 4.x version
Unfortunately, this still generates a lot of deprecation warnings from rails 5,
because this version adds strings to the rails middleware config, instead of constants.
https://github.com/airbrake/airbrake/blob/v4.3.8/lib/airbrake/railtie.rb#L17

I'm avoiding upgrading to 5.x now because that is a complete rewrite that
doesn't work properly with our forked version of errbit. We should upgrade
errbit or move to another service.