Remove ci_reporter_rspec gem from the project
The files outputted by this gem are no longer available in jenkins, and
even when they were, it's unclear that anyone was looking at them.  If
we remove this gem then our Jenkinsfile can be updated to use the standard
buildProject from govuk_jenkinslib.groovy and bring with it consistency
with other projects.