Use standard jenkins buildProject
We need to configure 3 things:

1. tell it not to sass lint (for now - we'll come back later and fix
   everything all in one go then turn this on)
2. tell it that our repo name is 'contacts-admin' because the jenkins job
   is just called 'contacts' and this would otherwise cause problems
3. tell it to set a 'RUNNING_IN_CI' env var in a beforeTest block.

The final change is because the standard buildProject will run db:reset to
prepare the db for running tests.  This runs db:seed and our db:seed tries
to import organisations from whitehall.  This won't work in a CI
environment, so we make seeds.rb check for the `RUNNING_IN_CI` environment
variable, and only run if it's not present.