Simplify redirecting gone contacts
Contacts Admin doesn't provide a way to redirect gone contacts, but
there is a rake task. It's pretty complicated though.

Simplify the rake task, to just take the content id of the "gone"
item, and the path to redirect to, and then unpublish the gone in the
Publishing API to setup the redirect.