Create form for deleting contact
Prior to this commit, clicking "Delete" triggered a pop-up with
the message "Are you sure?". The user would need to click "OK" to
confirm and delete the contact.

This commit removes the popup and creates a dedicated page instead.
The original "Delete" button now redirects to this page, which
asks the user to confirm to delete. Essentially it's the same UX.

This is to prepare for the next phase, which is to allow the user
to optionally specify a redirect URL when they delete the contact.
The presence of the redirect URL will determine whether we call
the existing `DestroyContact` interactor, or the
`DestroyAndRedirectContact` interactor (which is currently only
callable via a rake task and does not appear anywhere in the UI).