Validate `redirect_url` to only allow internal paths
Publishing API doesn't allow a `redirect_url` for external domains,
nor does it allow full GOV.UK URLs, returning a 422 Unprocessable
Entity.

The flash error wasn't rendering properly - I attributed this to
a non-existent `alert-error` class when the author intended to
use `alert-danger`.

It was also redirecting to the contacts index, regardless of if
the deletion succeeded: a stange user journey if the deletion
failed. This is now fixed up so that if the deletion fails, you
remain on the delete screen.

The problem now is that we ask for a 'redirect URL' when we really
want an internal path. In the next commit we'll allow full GOV.UK
URLs, following the precedent set in Whitehall.