Allow full GOV.UK URLs to be provided as redirects
Following the precedent set in Whitehall, we should be able to
pass full GOV.UK URLs to the publishing application, and rely on
the app to strip the domain out so that it only passed the path
to Publishing API.

I've made sure to test all possible variants of valid GOV.UK URLs.
The `remove_govuk_domain` could probably be implemented more
elegantly with a regex, but arguably at the expense of readability.

I've taken the opportunity to move some of the common testing
logic out into its own methods.