Revise and consolidate first part of the README
This is to be consistent with the pattern established in [1], with
just one set of instructions for running app commands.

I've removed a few pieces that I think are redundant:

- Technical intro. This is too vague to be useful, and can be learnt
from the API pages of the live examples.

- Tasks section. This can be easily seen in the code, and there isn't
enough detail to provide actionable insight.

- Note about timeouts. This can be an issue in many apps, and we should
document it centrally in GOV.UK Docker if it persists.

[1]: https://github.com/alphagov/content-publisher/pull/2257