Fix incorrect comparison in spec steps
The email address and phone number spec steps were comparing against the
`description` attribute. These models do not have such an attribute. It
appears the intention was to assert against the `title` attribute.

Updating this so the test works and we remove a warning which tells us
the test is pointless (as we are always comparing `nil` to `nil`).