Remove thin webserver dependency
This was used for running this app in development 8 years ago [1], it
can now safely use puma.

The motivation for this change is that I couldn't install thin on my M1
mac. Since the dependency was superfluous I didn't investigate whether
it was just a local setup issue or something more significant.

[1]: https://github.com/alphagov/contacts-admin/commit/3585748cb4d928d92047e52b3b2f67205b6bcba9