Remove Jenkinsfile special env var
We have this complex env var set-up so that the db:seed command doesn't
add data when in Jenkins. This is because the tests expect an empty
database. A simpler way to so this is to use the Rails.env value, since
we also want the same tests to pass locally.