Update paper_trail for Rails 7 compatibility
This stops the warning we get of:

```
PaperTrail 12.0.0 is not compatible with ActiveRecord 7.0.2.2. We allow PT
contributors to install incompatible versions of ActiveRecord, and this
warning can be silenced with an environment variable, but this is a bad
idea for normal use. Please install a compatible version of ActiveRecord
instead (>= 5.2, < 6.2). Please see the discussion in paper_trail/compatibility.rb
for details.
```