Add ActionSupport::Concern extension to Searchlight
When updating to Rails 7.1.2, the searchlight gem was producing a NoMethodError
(undefined method `class_attribute' for Searchlight::Adapters::ActionView:Module).
This occurred when requiring the gem in `value_objects/contacts_search.rb`. The only
way to resolve this was to duplicate the logic from Searchlight::Adapters::ActionView
and extend ActionSupport::Concern within it.