Remove rake task that relies on CSV input
Rake tasks that rely on CSV input are a throwback to the old
Puppet infrastructure of GOV.UK, where each app was running on
dedicated VMs, meaning there was always a fairly stable/reliable
file system to which we could upload said CSV and then run the
rake task against it via Jenkins. Kubernetes pods, in contrast,
are ephemeral in nature, so any uploaded CSVs could be lost at
any time. We should therefore migrate CSV-based rake tasks into
the UI of the respective publishing app, or remove it if it is no
longer needed.

This particular rake task was added 11 years ago in commit
d94d415ec7f2bdd38ba1592af271d5a25750fd4d. It was probably used to
do a one-off import of data, and I'm not aware of us having been
asked to run it again since, as publishers are able to add contacts
manually through the UI. It should therefore be safe to delete.

Trello: https://trello.com/c/yi1ExGvz/3179-audit-rake-tasks-that-rely-on-scp-push