Add ContactMigrator class ready for migrating contacts
Adds a basic test that the corresponding Contact (given a slug)
can be unpublished with a redirect.

Did consider calling this `MigrateContact` in line with the other
verb-driven name in this folder, `PublishFinders`. But looking
ahead to all the things we'll have to do in this migration, we're
going to want to migrate individual contacts _and_ we're going to
want to be able to kick off a batch of those (perhaps inside a
transaction). So naming the class `ContactMigrator` gives us the
freedom to extend its responsibility to all the other bits of
the migration we'll need to.