Wrap in a transaction to ensure every relevant Contact is migrated
The unpublish Publishing API calls are idempotent, so no risks
there. But what we DO risk is something going wrong mid task, and
then having deleted some of the Contact models, in which case the
task would then fail with our new `ContactNotFound` error.
We want them all to succeed, or none to succeed.